Refactor RPCEngine to use admin verification strategy
